Merged PR 727: [Feature] Improve dynamic location of Windows Kits executables.
Part of the capsule signing process has to locate tools that are installed with a Windows Driver Kit. These tools used to be located at a consistent path, but now the path has a dynamic component (version number).

This code addresses the problem in the Mu Basecore scripts while waiting for a common solution in the Mu Tools.